Rename test_upstream to highlight interactivity
