+++ src/mgr/hermes_test_manager/hermes_test_api.py
@@ -16,7 +16,7 @@ from test_cases import EnvironmentManager
 import test_cases.test_cases_dummy
 import test_cases.test_downstream_ifc
 import test_cases.test_downstream_ifc_interactive
-import test_cases.test_upstream_ifc
+import test_cases.test_upstream_ifc_interactive
 
 log = logging.getLogger('hermes_test_api')
