+++ src/mgr/hermes_test_manager/test_cases/test_upstream_ifc_interactive.py
